web_app/requirements.txt
streamlit
pandas
paho-mqtt
streamlit-lottie
